@@ -1,11 +1,11 @@
-import { browser, by, element } from 'protractor';
+import { browser, by, element } from "protractor";
 
 export class AppPage {
   navigateTo() {
-    return browser.get('/');
+    return browser.get("/");
   }
 
   getParagraphText() {
-    return element(by.css('app-root h1')).getText();
+    return element(by.css("app-root h1")).getText();
   }
 }
