@@ -1,9 +1,9 @@
-import { Component } from '@angular/core';
-import { AuthenticationService } from './authentication.service';
+import { Component } from "@angular/core";
+import { AuthenticationService } from "./authentication.service";
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html'
+  selector: "app-root",
+  templateUrl: "./app.component.html"
 })
 export class AppComponent {
   constructor(public auth: AuthenticationService) {}
