@@ -1,6 +1,6 @@
-import { Component } from '@angular/core';
+import { Component } from "@angular/core";
 
 @Component({
-  templateUrl: './home.component.html'
+  templateUrl: "./home.component.html"
 })
 export class HomeComponent {}
