@@ -1,23 +1,26 @@
-import { Component } from '@angular/core';
-import { AuthenticationService, TokenPayload } from '../authentication.service';
-import { Router } from '@angular/router';
+import { Component } from "@angular/core";
+import { AuthenticationService, TokenPayload } from "../authentication.service";
+import { Router } from "@angular/router";
 
 @Component({
-  templateUrl: './login.component.html'
+  templateUrl: "./login.component.html"
 })
 export class LoginComponent {
   credentials: TokenPayload = {
-    email: '',
-    password: ''
+    email: "",
+    password: ""
   };
 
   constructor(private auth: AuthenticationService, private router: Router) {}
 
   login() {
-    this.auth.login(this.credentials).subscribe(() => {
-      this.router.navigateByUrl('/profile');
-    }, (err) => {
-      console.error(err);
-    }); 
+    this.auth.login(this.credentials).subscribe(
+      () => {
+        this.router.navigateByUrl("/profile");
+      },
+      err => {
+        console.error(err);
+      }
+    );
   }
 }
