@@ -1,19 +1,22 @@
-import { Component } from '@angular/core';
-import { AuthenticationService, UserDetails } from '../authentication.service';
+import { Component } from "@angular/core";
+import { AuthenticationService, UserDetails } from "../authentication.service";
 
 @Component({
-  templateUrl: './profile.component.html'
+  templateUrl: "./profile.component.html"
 })
 export class ProfileComponent {
   details: UserDetails;
 
   constructor(private auth: AuthenticationService) {}
-  
-  ngOnInit() {    
-    this.auth.profile().subscribe(user => {
-      this.details = user;
-    }, (err) => {
-      console.error(err);
-    });
+
+  ngOnInit() {
+    this.auth.profile().subscribe(
+      user => {
+        this.details = user;
+      },
+      err => {
+        console.error(err);
+      }
+    );
   }
 }
