@@ -1,24 +1,27 @@
-import { Component } from '@angular/core';
-import { AuthenticationService, TokenPayload } from '../authentication.service';
-import { Router } from '@angular/router';
+import { Component } from "@angular/core";
+import { AuthenticationService, TokenPayload } from "../authentication.service";
+import { Router } from "@angular/router";
 
 @Component({
-  templateUrl: './register.component.html'
+  templateUrl: "./register.component.html"
 })
 export class RegisterComponent {
   credentials: TokenPayload = {
-    email: '',
-    name: '',
-    password: ''
+    email: "",
+    name: "",
+    password: ""
   };
 
   constructor(private auth: AuthenticationService, private router: Router) {}
 
   register() {
-    this.auth.register(this.credentials).subscribe(() => {
-      this.router.navigateByUrl('/profile');
-    }, (err) => {
-      console.error(err);
-    });
+    this.auth.register(this.credentials).subscribe(
+      () => {
+        this.router.navigateByUrl("/profile");
+      },
+      err => {
+        console.error(err);
+      }
+    );
   }
 }
